Add error_page middleware twice
The first time we add the middleware needs to be before the message and
auth layers so that the user and messages can be displayed in the error
page. This is more necessary for the user object, otherwise it would be
quite jarring to see yourself logged out in the header when you went to
a not found page.

However, adding it twice seems a bit hacky. Potentially the second
instance should just be in the axum .fallback() method. Need to look
more into that.